Moved tests to own package
This stops them being packaged up for PyPI where they are not needed.
Hence django-adaptors is smaller and quicker to install.

I removed some more unnecessary files.
